fix: (platform) apply comments after review
