+++ apps/docs/src/app/platform/component-docs/platform-combobox/examples/combobox-columns/combobox-columns-example.component.html
@@ -9,6 +9,6 @@
             secondaryTextAlignment="right"
             (selectionChange)="onSelect($event)"
         ></fdp-combobox>
-        <p>Selected Item: {{ selectedItem }}</p>
+        <p>Selected Item: {{ selectedItem | json }}</p>
     </fdp-form-field>
 </fdp-form-group>
