+++ apps/docs/src/app/platform/component-docs/platform-combobox/examples/combobox-group/combobox-group-example.component.html
@@ -8,7 +8,7 @@
             groupKey="type"
             (selectionChange)="onSelect($event)"
         ></fdp-combobox>
-        <p>Selected: {{ selectedItem }}</p>
+        <p>Selected: {{ selectedItem | json }}</p>
     </fdp-form-field>
 
     <fdp-form-field id="group-without-key" placeholder="Type some text..." label="Group without Key" zone="zLeft"
@@ -20,6 +20,6 @@
             group="true"
             (selectionChange)="onSelect1($event)"
         ></fdp-combobox>
-        <p>Selected: {{ selectedItem1 }}</p>
+        <p>Selected: {{ selectedItem1 | json }}</p>
     </fdp-form-field>
 </fdp-form-group>
