+++ apps/docs/src/app/platform/component-docs/platform-combobox/examples/combobox-mobile/combobox-mobile-example.component.html
@@ -8,6 +8,6 @@
             [mobileConfig]="mobileConfig"
             (selectionChange)="onSelect($event)"
         ></fdp-combobox>
-        <p>Selected: {{ selectedItem }}</p>
+        <p>Selected: {{ selectedItem | json }}</p>
     </fdp-form-field>
 </fdp-form-group>
