+++ libs/platform/src/lib/components/form/combobox/combobox/combobox.component.scss
@@ -1,7 +1,7 @@
 @import '~fundamental-styles/dist/list.css';
 
 .fd-select-overlay {
-    width: 100%;
+    width: 100% !important;
 }
 
 .fd-combobox-input-menu-overflow {
