+++ libs/platform/src/lib/components/form/combobox/combobox/combobox.component.ts
@@ -1,17 +1,11 @@
 import {
-    AfterViewInit,
     ChangeDetectionStrategy,
     ChangeDetectorRef,
     Component,
-    ContentChildren,
     ElementRef,
-    EventEmitter,
     Inject,
-    Input,
     OnInit,
     Optional,
-    Output,
-    QueryList,
     Self,
     TemplateRef,
     ViewChild,
@@ -23,12 +17,10 @@ import {
     DIALOG_CONFIG,
     DialogConfig,
     DynamicComponentService,
-    MenuKeyboardService,
-    MobileModeConfig,
-    TemplateDirective
+    MenuKeyboardService
 } from '@fundamental-ngx/core';
 import { ComboBoxDataSource, DATA_PROVIDERS, DataProvider } from '../../../../domain/data-source';
-import { ContentDensity, FormFieldControl } from '../../form-control';
+import { FormFieldControl } from '../../form-control';
 import { BaseCombobox, ComboboxSelectionChangeEvent } from '../commons/base-combobox';
 import { ComboboxConfig } from '../combobox.config';
 import { OptionItem } from '../../../../domain';
@@ -41,68 +33,7 @@ import { OptionItem } from '../../../../domain';
     encapsulation: ViewEncapsulation.None,
     providers: [{ provide: FormFieldControl, useExisting: ComboboxComponent, multi: true }]
 })
-export class ComboboxComponent extends BaseCombobox<string> implements OnInit, AfterViewInit {
-    /** Whether the autocomplete should be enabled; Enabled by default */
-    @Input()
-    autoComplete = true;
-
-    /**
-     * content Density of element. 'cozy' | 'compact'
-     */
-    @Input()
-    set contentDensity(contentDensity: ContentDensity) {
-        this._contentDensity = contentDensity;
-        this.isCompact = contentDensity === 'compact';
-    }
-
-    /**
-     * Todo: Name of the entity for which DataProvider will be loaded. You can either pass list of
-     * items or use this entityClass and internally we should be able to do lookup to some registry
-     * and retrieve the best matching DataProvider that is set on application level
-     */
-    @Input()
-    entityClass: string;
-
-    /** Whether the combobox is readonly. */
-    @Input()
-    readOnly = false;
-
-    /** Whether the combobox should be built on mobile mode */
-    @Input()
-    mobile = false;
-
-    /** Multi Input Mobile Configuration, it's applied only, when mobile is enabled */
-    @Input()
-    mobileConfig: MobileModeConfig;
-
-    /** Tells the combo if we need to group items */
-    @Input()
-    group = false;
-
-    /** A field name to use to group data by (support dotted notation) */
-    @Input()
-    groupKey?: string;
-
-    /** The field to show data in secondary column */
-    @Input()
-    secondaryKey?: string;
-
-    /** Show the second column (Applicable for two columns layout) */
-    @Input()
-    showSecondaryText = false;
-
-    /** Horizontally align text inside the second column (Applicable for two columns layout) */
-    @Input()
-    secondaryTextAlignment: 'left' | 'center' | 'right' = 'right';
-
-    /** Turns on/off Adjustable Width feature */
-    @Input()
-    autoResize = false;
-
-    /** @hidden */
-    @ContentChildren(TemplateDirective)
-    customTemplates: QueryList<TemplateDirective>;
-
+export class ComboboxComponent extends BaseCombobox implements OnInit {
     /** @hidden */
     @ViewChild('searchInputElement')
     searchInputElement: ElementRef;
@@ -116,31 +47,7 @@ export class ComboboxComponent extends BaseCombobox<string> implements OnInit, A
     listTemplate: TemplateRef<any>;
 
     /** @hidden */
-    optionItemTemplate: TemplateRef<any>;
-
-    /** @hidden */
-    groupItemTemplate: TemplateRef<any>;
-
-    /** @hidden */
-    secondaryItemTemplate: TemplateRef<any>;
-
-    /** @hidden */
-    selectedItemTemplate: TemplateRef<any>;
-
-    @Output()
-    selectionChange = new EventEmitter<ComboboxSelectionChangeEvent>();
-
-    /** @hidden */
-    _contentDensity: ContentDensity = this._comboboxConfig.contentDensity;
-
-    /**
-     * @hidden
-     * Whether "contentDensity" is "compact"
-     */
-    isCompact: boolean = this._contentDensity === 'compact';
-
-    /** @hidden */
-    protected selected = '';
+    selected = '';
 
     get isGroup(): boolean {
         return !!(this.group && this.groupKey);
@@ -169,12 +76,6 @@ export class ComboboxComponent extends BaseCombobox<string> implements OnInit, A
         }
     }
 
-    /** @hidden */
-    ngAfterViewInit(): void {
-        this._assignCustomTemplates();
-        super.ngAfterViewInit();
-    }
-
     /** @hidden
      * Method to emit change event
      */
@@ -198,24 +99,24 @@ export class ComboboxComponent extends BaseCombobox<string> implements OnInit, A
     /** @hidden
      * Method to set as selected
      */
-    setAsSelected(item: OptionItem[]): void {
-        const optionItem = item[0];
+    setAsSelected(item: any[]): void {
+        const selectedItem = item[0];
 
-        if (this.isSelectedOptionItem(optionItem)) {
+        if (this.isSelectedOptionItem(selectedItem)) {
             return;
         }
 
-        this.inputTextValue = this.displayValue(optionItem);
-        this.selected = this.lookupValue(optionItem) || this.inputTextValue;
+        this.inputTextValue = this.displayValue(selectedItem);
+        this.selected = this.lookupValue(selectedItem) || this.inputTextValue;
 
-        this._updateModel(optionItem.value);
+        this._updateModel(selectedItem);
     }
 
     /** @hidden
      * Define is selected item selected
      */
-    isSelectedOptionItem(selectedItem: OptionItem): boolean {
-        return this.selected === this.lookupValue(selectedItem) || this.selected === selectedItem.label;
+    isSelectedOptionItem(selectedItem: any): boolean {
+        return this.selected === this.lookupValue(selectedItem) || this.selected === this.displayValue(selectedItem);
     }
 
     /** @hidden
@@ -237,27 +138,5 @@ export class ComboboxComponent extends BaseCombobox<string> implements OnInit, A
         this.value = value;
 
         this.emitChangeEvent(value);
-    }
-
-    /** @hidden
-     * Assign custom templates
-     * */
-    private _assignCustomTemplates(): void {
-        this.customTemplates.forEach(template => {
-            switch (template.getName()) {
-                case 'optionItemTemplate':
-                    this.optionItemTemplate = template.templateRef;
-                    break;
-                case 'groupItemTemplate':
-                    this.groupItemTemplate = template.templateRef;
-                    break;
-                case 'secondaryItemTemplate':
-                    this.secondaryItemTemplate = template.templateRef;
-                    break;
-                case 'selectedItemTemplate':
-                    this.selectedItemTemplate = template.templateRef;
-                    break;
-            }
-        });
     }
 }
